Passes: some passes moved to node class
